% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Metadata_bib.R
\name{add_metadata_bib}
\alias{add_metadata_bib}
\title{Add bibliographic metadata to a \code{ZenodoRecord}}
\usage{
add_metadata_bib(
  metadata_bib = new_metadata_bib(),
  rec = zen4R::ZenodoRecord$new()
)
}
\arguments{
\item{metadata_bib}{an object of class \code{metadata_bib} as created by the
function \code{new_metadata_bib()}.}

\item{rec}{an object of class \code{ZenodoRecord} s created by
\code{zen4R::ZenodoRecord$new()}. If missing, a new one will be created.}
}
\value{
the object \code{rec} with the metadata in \code{metdata_bib} added.
}
\description{
Add bibliographic metadata to a \code{ZenodoRecord}
}
