% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leef_create_data_archives_per_method.R
\name{leef_create_data_archives_per_method}
\alias{leef_create_data_archives_per_method}
\title{Create the data packages from the LEEF project.}
\usage{
leef_create_data_archives_per_method(
  to_dir = ".",
  data_dir = "/Volumes/LEEF-1_archive/LEEF.archived.data/LEEF/3.archived.data",
  metadata_dir = "/Volumes/LEEF/LEEF-1.metadata",
  stage = c("pre_processed", "extracted"),
  cores = 1
)
}
\arguments{
\item{to_dir}{directory in which the compressed data folders should be saved to}

\item{data_dir}{base directory of the data archive}

\item{stage}{stage of the data. Allowed values are \code{"pre_processed"},
\code{"extracted"} or \code{c("pre_processed", "extracted")}}

\item{cores}{number of cores to be used for the parallel compression (default
is one due to limits in connection)}
}
\value{
names of the created data zip archives
}
\description{
The data packages contain only of the respective data, bundled in a zipped file.
}
