% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leef_deposit.R
\name{leef_deposit}
\alias{leef_deposit}
\title{Deposit data on Zenodo}
\usage{
leef_deposit(
  timestamp,
  token = zen4R:::zenodo_pat(),
  data_package_dir,
  metadata_extracted,
  metadata_pre_processed,
  metadata_bib,
  publish = FALSE,
  sandbox = TRUE
)
}
\arguments{
\item{timestamp}{timestamp to be uploaded}

\item{token}{Zenodo token to upload the deposit, The default is to take it
from the environment variable \code{ZENODO_PAT}. The internal helper function
from \code{zen4R} is used. You can set it using \code{Sys.setenv(ZENODO_PAT = "YOUR_PAT_HERE")}.}

\item{data_package_dir}{directory in which the datapackages to be deposited can be found}

\item{metadata_extracted}{file name of the zip file with the metadata for the extracted data}

\item{metadata_pre_processed}{file name of the zip file with the metadata for the pre-processed data}

\item{metadata_bib}{bibliographic metadata as generated by the function
\code{zen_metadata()}}

\item{publish}{if \code{TRUE} publish the packages immediately - should
normally be \code{FALSE}}

\item{sandbox}{if \code{TRUE} (default) upload to the Zenodo sandbox for testing,
if \code{FALSE} upload to the "real" Zenodo}
}
\value{
list with all info concerning the deposits
}
\description{
Deposit data on Zenodo
}
