% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leef_metadata_bib.R
\name{leef_metadata_bib}
\alias{leef_metadata_bib}
\title{Create list containing the bibligraphic metadata for the data deposit}
\usage{
leef_metadata_bib(
  community = "leef-uzh",
  upload_type = "dataset",
  authors = list(RMK = list(firstname = "Rainer M", lastname = "Krug", affiliation =
    "University of Zurich", orcid = "0000-0002-7490-0066"), OLP = list(firstname =
    "Owen L", lastname = "Petchey", affiliation = "University of Zurich", orcid =
    "0000-0002-7724-1633")),
  description = "Description of the data",
  version = "1.0.0",
  language = "eng",
  keywords = c("LEEZ-UZH", "LEEF-1"),
  access_right = "open",
  license = "CC-BY-SA-4.0",
  contributors = list(RMK = list(firstname = "Rainer M", lastname = "Krug", type =
    "DataManager", affiliation = "University of Zurich", orcid = "0000-0002-7490-0066"),
    OLP = list(firstname = "Owen L", lastname = "Petchey", type = "ProjectLeader",
    affiliation = "University of Zurich", orcid = "0000-0002-7724-1633"))
)
}
\arguments{
\item{community}{community to which the deposited data should be added}

\item{upload_type}{type of the deposited data}

\item{authors}{authors of the deposited data. Each contributor
is also a list with the following fields:
\itemize{
\item firstname
\item lastname
\item affiliation
\item orcid
}}

\item{description}{description of the deposited data}

\item{version}{version of the deposited data}

\item{language}{language of the deposited data}

\item{keywords}{keywords}

\item{access_right}{access right of the deposited data}

\item{license}{license of the deposited data}

\item{contributors}{list of contributors of the deposited data. Each contributor
is also a list with the following fields:
\itemize{
\item firstname
\item lastname
\item type
\item affiliation
\item orcid
}}
}
\value{
list containing the metadata necessary for the data deposit
}
\description{
Create list containing the bibligraphic metadata for the data deposit
}
