% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_metadata_bib.R
\name{new_metadata_bib}
\alias{new_metadata_bib}
\title{Create list containing the bibligraphic metadata for the data deposit}
\usage{
new_metadata_bib(
  title = "Some Title",
  description = "Description of the data",
  keywords = c("LEEZ-UZH", "LEEF-1"),
  community = "leef-uzh",
  upload_type = "dataset",
  authors = list(RMK = list(firstname = "Rainer M", lastname = "Krug", affiliation =
    "University of Zurich", orcid = "0000-0002-7490-0066"), OLP = list(firstname =
    "Owen L", lastname = "Petchey", affiliation = "University of Zurich", orcid =
    "0000-0002-7724-1633")),
  version = "1.0.0",
  language = "eng",
  access_right = "embargoed",
  embargo_date = as.Date("2030-01-13"),
  license = "CC-BY-SA-4.0",
  contributors = list(RMK = list(firstname = "Rainer M", lastname = "Krug", type =
    c("DataManager", "Software"), affiliation = "University of Zurich", orcid =
    "0000-0002-7490-0066"), OLP = list(firstname = "Owen L", lastname = "Petchey", type =
    "ProjectLeader", affiliation = "University of Zurich", orcid =
    "0000-0002-7724-1633")),
  grants = "some names of the grant"
)
}
\arguments{
\item{title}{title of the deposit}

\item{description}{description of the deposited data}

\item{keywords}{\code{character} vector with keywords}

\item{upload_type}{type of the deposited data}

\item{authors}{authors of the deposited data. Each contributor is also a list
with the following fields:
\itemize{
\item firstname
\item lastname
\item affiliation
\item orcid
}}

\item{version}{version of the deposited data}

\item{language}{language of the deposited data}

\item{access_right}{access right of the deposited data}

\item{license}{license of the deposited data}

\item{contributors}{list of contributors of the deposited data. Each
contributor is also a list with the following fields:
\itemize{
\item firstname
\item lastname
\item type: the role. This \code{character} vector can have multiple values which will
be expanded for submission to Zenodo
\item affiliation
\item orcid
}}

\item{communities}{\code{character} vector containing the community idsto which
the deposited data should be added to,}

\item{embargo_data}{date until which the data is embargoed if \code{access_right}
is "embargoed"}
}
\value{
object of class \code{metadata_bib} and \code{list}. It is used by the function
\code{add_metadata_bib()} to be added to the record which will be deposited to
Zenodo.
}
\description{
Create list containing the bibligraphic metadata for the data deposit
}
