% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_metadata_bib.R
\name{read_metadata_bib}
\alias{read_metadata_bib}
\title{Read metadata_bib object from yaml file}
\usage{
read_metadata_bib(yamlfile)
}
\arguments{
\item{yamlfile}{name of the file}

\item{metadata_bib}{object of class `metadata_bib`}
}
\value{
result from `yaml::write_yaml()`
}
\description{
A essentially simple wrapper around `yaml::write_yaml()`
}
