% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_metadata_bib.R
\name{save_metadata_bib}
\alias{save_metadata_bib}
\title{Save metadata_bib object to yaml file}
\usage{
save_metadata_bib(metadata_bib, yamlfile)
}
\arguments{
\item{metadata_bib}{object of class `metadata_bib`}

\item{yamlfile}{name of the file}
}
\value{
result from `yaml::write_yaml()`
}
\description{
A simple wrapper around `yaml::write_yaml()`
}
