% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen_create_data_archives.R
\name{zen_create_data_archives}
\alias{zen_create_data_archives}
\title{Title}
\usage{
zen_create_data_archives(
  to_dir = ".",
  archive_dir = "~/Duck/LEEFSwift3",
  stage,
  cores = 4
)
}
\arguments{
\item{to_dir}{directory in which the compressed data folders should be saved to}

\item{archive_dir}{base directory of the data archive}

\item{stage}{stage of the data. Allowed values are \code{"pre_processed"}, \code{"extracted"} or \code{c("pre_processed", "extracted"}}

\item{cores}{number of cores to be used for the parallel compression}
}
\value{
names of the created data zip archives
}
\description{
Title
}
