% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen_deposit.R
\name{zen_deposit}
\alias{zen_deposit}
\title{Deposit data on Zenodo}
\usage{
zen_deposit(
  timestamp,
  token = NULL,
  data_dir = "./tmp",
  metadata_bib = leef_metadata_bib(),
  publish = FALSE,
  sandbox = TRUE
)
}
\arguments{
\item{timestamp}{timestamp to be uploaded}

\item{token}{Zenodo token to upload the deposit}

\item{data_dir}{directory in which the data to be deposited can be found}

\item{metadata_bib}{bibliographic metadata as generated by the function \code{zen_metadata()}}

\item{publish}{if \code{TRUE} publish the packages immediately - should normally be \code{FALSE}}

\item{sandbox}{if \code{TRUE} (default) upload to the Zenodo sandbox for testing, if \code{FALSE} upload to the "real" Zenodo}
}
\value{
list with all info concerning the deposits
}
\description{
Deposit data on Zenodo
}
