% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leef_create_data_archives.R
\name{leef_create_data_archives}
\alias{leef_create_data_archives}
\title{Title}
\usage{
leef_create_data_archives(
  to_dir = ".",
  archive_dir = "~/Duck/LEEFSwift3",
  timestamp,
  stage = c("pre_processed", "extracted")
)
}
\arguments{
\item{to_dir}{directory in which the compressed data folders should be saved to}

\item{archive_dir}{base directory of the data archive}

\item{timestamp}{timestamp to be uploaded}

\item{stage}{stage of the data. Allowed values are \code{"pre_processed"}, \code{"extracted"}}
}
\value{
names of the created data zip archives
}
\description{
Title
}
