% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_metadata_bib.R
\name{new_metadata_bib}
\alias{new_metadata_bib}
\title{Create list containing the bibligraphic metadata for the data deposit}
\usage{
new_metadata_bib(
  title = "Some Title",
  description = "Description of the data",
  subject = list(a = "LEEZ-UZH", b = "LEEF-1"),
  creator = list(RMK = list(name = "Rainer M Krug", affiliation = "University of Zurich",
    orcid = "0000-0002-7490-0066"), OLP = list(name = "Owen L Petchey", affiliation =
    "University of Zurich", orcid = "0000-0002-7724-1633")),
  language = "eng",
  accessRights = "embargoed",
  license = "CC-BY-SA-4.0",
  contributor = list(RMK = list(name = "Rainer M Krug", type = c("DataManager",
    "Software"), affiliation = "University of Zurich", orcid = "0000-0002-7490-0066"),
    OLP = list(name = "Owen L Petchey", type = "ProjectLeader", affiliation =
    "University of Zurich", orcid = "0000-0002-7724-1633")),
  isPartOf = NULL
)
}
\arguments{
\item{title}{title of the deposit}

\item{description}{description of the deposited data}

\item{subject}{\code{character} vector with keywords
the deposited data should be added to,}

\item{language}{language of the deposited data}

\item{accessRights}{access right of the deposited data}

\item{license}{license of the deposited data}

\item{contributor}{list of contributors of the deposited data. Each
contributor is also a list with the following fields:
\itemize{
\item firstname
\item lastname
\item type: the role. This \code{character} vector can have multiple values which will
be expanded for submission to Zenodo
\item affiliation
\item orcid
}}

\item{embargo_data}{date until which the data is embargoed if \code{accessRights}
is "embargoed"}
}
\value{
object of class \code{metadata_bib} and \code{list}. It is used by the function
\code{add_metadata_bib()} to be added to the record which will be deposited to
Zenodo.
}
\description{
Create list containing the bibligraphic metadata for the data deposit
}
