% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen_deposit.R
\name{zen_deposit}
\alias{zen_deposit}
\title{Deposit data on Zenodo}
\usage{
zen_deposit(
  token = NULL,
  timestamp,
  data_dir = "./tmp",
  metadata_bib = zen_metadata_bib(),
  publish = FALSE,
  sandbox = TRUE
)
}
\arguments{
\item{token}{Zenodo token to upload the deposit}

\item{timestamp}{timestamp to be uploaded}

\item{publish}{if \code{TRUE} publish the packages immediately - should normally be \code{FALSE}}

\item{sandbox}{if \code{TRUE} (default) upload to the Zenodo sandbox for testing, if \code{FALSE} upload to the "real" Zenodo}

\item{measuring_method}{measuring method to be uploaded. Allowed values at the moment:
\itemize{
\item \code{"bemovi.mag.16"}
\item \code{"bemovi.mag.25"}
\item \code{"flowcam"}
\item \code{"flowcytometer"}
\item \code{"o2meter"}
\item \code{"manualcount"}
}}

\item{metadata}{metadata as genersted by the function \code{zen_metadata()}}

\item{archive_dir}{base directory of the data archive}
}
\value{
list with all info concerning the deposits
}
\description{
Deposit data on Zenodo
}
